restructure UI
