+++ app/src/androidTest/java/com/example/pokergameui/ProtocolTest.kt
@@ -1,10 +1,8 @@
 package com.example.pokergameui
 
-import android.util.Log
-import com.daveanthonythomas.moshipack.MoshiPack
+import com.example.pokergameui.model.LoginRequest
+import com.example.pokergameui.model.Protocol
 import org.junit.Test
-import java.nio.ByteBuffer
-import java.nio.ByteOrder
 
 class ProtocolTest {
     @Test
@@ -13,10 +11,9 @@ class ProtocolTest {
         assert(encodedPacket != null)
         if (encodedPacket != null) {
             val decodedPacket = Protocol.decode<LoginRequest>(encodedPacket.array())
-            assert(decodedPacket != null)
-            assert(decodedPacket?.header?.packetType == 100.toShort())
-            assert(decodedPacket?.payload?.user == "admin")
-            assert(decodedPacket?.payload?.pass == "admin")
+            assert(decodedPacket.header.packetType == 100.toShort())
+            assert(decodedPacket.payload?.user == "admin")
+            assert(decodedPacket.payload?.pass == "admin")
         }
     }
 }
