+++ app/src/main/AndroidManifest.xml
@@ -14,58 +14,30 @@
         android:theme="@style/Theme.PokerGameUI"
         tools:targetApi="31">
        <activity
-            android:name=".StartActivity"
-            android:screenOrientation="userLandscape"
+            android:name=".view.StartActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.PokerGameUI"
             >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
 
         <activity
-            android:name=".SignUpActivity"
-            android:screenOrientation="userPortrait"
-            android:exported="true"
-            android:label="@string/app_name"
-            android:theme="@style/Theme.PokerGameUI"
-            >
-
-        </activity>
-        <activity
-            android:name=".SignInActivity"
-            android:screenOrientation="userPortrait"
-            android:exported="true"
-            android:label="@string/app_name"
-            android:theme="@style/Theme.PokerGameUI"
-            >
-        </activity>
-        <activity
-            android:name=".InGameActivity"
-            android:screenOrientation="landscape"
+            android:name=".view.InGameActivity"
+            android:screenOrientation="userLandscape"
             android:exported="true"
             android:theme="@style/Theme.PokerGameUI"
             >
         </activity>
         <activity
-            android:name=".LobbyActivity"
+            android:name=".view.LobbyActivity"
             android:exported="true"
             android:theme="@style/Theme.PokerGameUI"
             >
         </activity>
-        <activity
-            android:name=".JoinActivity"
-            android:exported="true"
-            android:theme="@style/Theme.PokerGameUI"/>
-        <activity
-            android:name=".CreateActivity"
-            android:exported="true"
-            android:theme="@style/Theme.PokerGameUI"/>
-
 
     </application>
