+++ app/src/main/java/com/example/pokergameui/view/SignInScreen.kt
@@ -1,0 +1,181 @@
+package com.example.pokergameui.view
+
+import android.content.Context
+import android.content.Intent
+import android.util.Log
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.input.TextFieldValue
+import androidx.compose.ui.text.style.TextDecoration
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.pokergameui.ui.theme.Blue
+import com.example.pokergameui.ui.theme.Dark
+import com.example.pokergameui.ui.theme.InputLabel
+import com.example.pokergameui.viewmodel.SigninViewModel
+
+fun handleSignIn(context: Context, username: String, password: String) {
+    if (username.isEmpty() || password.isEmpty()) {
+        Log.d("SignInScreen", "Username or password is empty")
+        return
+    }
+
+    SigninViewModel().signIn(username, password)
+
+    val intent = Intent(context, LobbyActivity::class.java)
+    context.startActivity(intent)
+}
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun SignInScreen(onBackPressed: () -> Unit) {
+    Scaffold(
+        topBar = {
+            TopAppBar(
+                title = { Text("Sign In") },
+                navigationIcon = {
+                    IconButton(onClick = onBackPressed) {
+                        Icon(Icons.AutoMirrored.Filled.ArrowBack, contentDescription = "Back")
+                    }
+                }
+            )
+        }
+    ) { innerPadding ->
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(innerPadding)
+                .padding(horizontal = 16.dp),
+            verticalArrangement = Arrangement.Center
+        ) {
+            SignInHeader()
+            Spacer(modifier = Modifier.height(24.dp))
+            SignInForm(modifier = Modifier.fillMaxWidth(), handleSignIn = ::handleSignIn)
+        }
+    }
+}
+
+
+@Composable
+fun SignInHeader(modifier: Modifier = Modifier) {
+    Surface(color = Dark, contentColor = Color.White) {
+        Column(
+            modifier = modifier
+                .fillMaxWidth()
+                .padding(vertical = 24.dp, horizontal = 16.dp),
+            verticalArrangement = Arrangement.Center
+        ) {
+            Text(
+                text = "Sign in to your Account",
+                fontSize = 28.sp,
+                lineHeight = 36.sp,
+                fontWeight = FontWeight.Bold
+            )
+            Spacer(modifier = Modifier.height(8.dp))
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                horizontalArrangement = Arrangement.Start
+            ) {
+                Text(
+                    text = "Don't have an account? ",
+                    fontSize = 14.sp,
+                    color = Color.White
+                )
+                Text(
+                    text = "Sign up",
+                    color = Blue,
+                    textDecoration = TextDecoration.Underline,
+                    fontSize = 14.sp,
+                    modifier = Modifier.clickable {
+                        // Navigate to Sign-Up screen logic
+
+                    }
+                )
+            }
+        }
+    }
+}
+
+@Composable
+fun SignInForm(
+    modifier: Modifier = Modifier,
+    handleSignIn: (Context, String, String) -> Unit = { _, _, _ -> },
+    handleForgetPassWord: () -> Unit = {}
+) {
+    var username by remember { mutableStateOf(TextFieldValue("")) }
+    var password by remember { mutableStateOf(TextFieldValue("")) }
+    val context = LocalContext.current
+
+    Column(
+        modifier = modifier
+            .padding(vertical = 16.dp)
+    ) {
+        InputLabel(
+            onValueChange = { username = it },
+            value = username,
+            label = "Email",
+            placeholder = "Enter your email",
+            modifier = Modifier.fillMaxWidth()
+        )
+        Spacer(modifier = Modifier.height(12.dp))
+        InputLabel(
+            label = "Password",
+            value = password,
+            onValueChange = { password = it },
+            placeholder = "Enter your password",
+            hidden = true,
+            modifier = Modifier.fillMaxWidth()
+        )
+        Spacer(modifier = Modifier.height(16.dp))
+        Button(
+            colors = ButtonDefaults.buttonColors(containerColor = Blue),
+            onClick = { handleSignIn(context, username.text, password.text) },
+            modifier = Modifier.fillMaxWidth()
+        ) {
+            Text(
+                text = "Login",
+                fontSize = 18.sp,
+                fontWeight = FontWeight.SemiBold
+            )
+        }
+        Spacer(modifier = Modifier.height(12.dp))
+        Text(
+            text = "Forgot Password?",
+            color = Blue,
+            textDecoration = TextDecoration.Underline,
+            modifier = Modifier
+                .align(Alignment.CenterHorizontally)
+                .clickable {
+                    handleForgetPassWord()
+                }
+        )
+    }
+}
